+++ src/components/Header/index.js
@@ -28,25 +28,42 @@ const Title = styled.h3`
   margin: 0;
 `;
 
+const MenuWrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
+
 type HeaderProps = {
   title?: string,
   userMenuProps?: UserMenuProps,
+  additionalUserMenuProps?: UserMenuProps,
   tabNavigationProps?: TabNavigationProps,
   className?: string,
   style?: Object,
 };
 
-const Header = ({ title, userMenuProps, tabNavigationProps, className, style }: HeaderProps) => (
+const Header = ({
+  title,
+  userMenuProps,
+  additionalUserMenuProps,
+  tabNavigationProps,
+  className,
+  style,
+}: HeaderProps) => (
   <Container className={className} style={style}>
     {title && <Title>{title}</Title>}
     {tabNavigationProps && <TabNavigation {...tabNavigationProps} />}
-    {userMenuProps && <UserMenu {...userMenuProps} />}
+    <MenuWrapper>
+      {additionalUserMenuProps && <UserMenu {...additionalUserMenuProps} />}
+      {userMenuProps && <UserMenu {...userMenuProps} />}
+    </MenuWrapper>
   </Container>
 );
 
 Header.defaultProps = {
   title: null,
   userMenuProps: null,
+  additionalUserMenuProps: null,
   tabNavigationProps: null,
   className: '',
   style: {},